feat: WebQSP 数据加载健壮性 + AMP 混合精度训练支持
- data.py: 新增 normalize_answers、缺失实体跳过并警告、pin_memory、使用 qa_test_webqsp_fixed.txt
- model.py: follow() 中将 r 强制转为 float32，兼容 AMP 稀疏矩阵乘法
- train.py: 引入 autocast + GradScaler，clip_grad_value_ 改为 clip_grad_norm_ 统一梯度裁剪

+++ WebQSP/model.py
@@ -41,7 +41,8 @@ class TransferNet(nn.Module):
 
 
     def follow(self, e, r):
-        x = torch.sparse.mm(self.Msubj, e.t()) * torch.sparse.mm(self.Mrel, r.t())
+        # cast r to float32: sparse mm doesn't support float16 (AMP compat)
+        x = torch.sparse.mm(self.Msubj, e.t()) * torch.sparse.mm(self.Mrel, r.float().t())
         return torch.sparse.mm(self.Mobj.t(), x).t() # [bsz, Esize]
 
     def forward(self, heads, questions, answers=None, entity_range=None):
